Delete BasicTextarea component
