--- frontend/src/components/BasicTextarea/index.ts
@@ -1,1 +1,0 @@
-export * from './BasicTextarea';
